+++ src/app/dashboard/artifact/artifact.component.ts
@@ -28,7 +28,7 @@ import { TableComponent } from './table/table.component'
     selector: 'app-artifact',
     templateUrl: './artifact.component.html',
     styleUrls: ['./artifact.component.scss'],
-    imports: [CommonModule, MatExpansionModule, NgScrollbarModule, MarkdownComponent, TranslocoModule],
+    imports: [CommonModule, MatExpansionModule, NgScrollbarModule, TranslocoModule],
     encapsulation: ViewEncapsulation.None
 })
 export class ArtifactComponent implements OnInit, AfterViewInit {
@@ -101,6 +101,7 @@ export class ArtifactComponent implements OnInit, AfterViewInit {
 
     ngOnInit(): void {
         if (this.artifact) {
+            this.modality = this.artifact.modality
             this.getService().fetchArtifact(this.artifact)
         }